pyproject.toml
[tool.poetry]
name = "secretgraph"
description = "A secure, decentral identity graph"
version = "0.1"
license = "MIT"
authors = ["Alexander Kaftan"]
homepage = "https://secretgraph.net"
readme = "README.md"
repository = "https://github.com/devkral/secretgraph"
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]

keywords = [
  "online identity", "decentral networks"
]

[tool.poetry.dependencies]
python = ">=3.9, <4"
cryptography = "*"
rdflib = '*'
certifi = "*"
httpx = {version = "*", extras = ["http2"]}
gql = {version = "*", extras = ["aiohttp", "websockets"]}
django = {version = '>=4.2', optional = true}
channels = {version = ">=4.0", optional = true}
graphene-protector = {version=">=0.8.0", optional = true}
strawberry-graphql = {version=">=0.183",  optional = true}
strawberry-django-plus = {version="*", optional = true}
flipflop = {version="*", optional = true}
mysqlclient = {version="*", optional = true}
psycopg = {version="*", extras=["binary","pool"], optional = true}
django-manifest-loader = {version="*", optional = true}
django-fast-ratelimit = {version=">=0.9.0", optional = true}
bleach = {version=">=6.0", optional = true}
tinycss2 = {version="*", optional = true}


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
django-webtest = "*"
WSGIProxy2 = "*"
django-manifest-loader = "*"
django-fast-ratelimit = ">=0.9.0"
bleach = "*"
tinycss2 = "*"
graphene-protector = ">=0.7.2"
django = ">=4.1"
channels = {version = '>=4.0', extras = ["daphne"] }
strawberry-graphql = {version = '>=0.145', extras = ["debug"] }
strawberry-django-plus = "*"
tox = "*"


[tool.poetry.group.dev.dependencies]
django-manifest-loader = "*"
django-fast-ratelimit = ">=0.9.0"
bleach = "*"
tinycss2 = "*"
graphene-protector = ">=0.7.2"
django = ">=4.2"
channels = {version = '>=4.0', extras = ["daphne"] }
strawberry-graphql = {version = '>=0.145', extras = ["debug"] }
strawberry-django-plus = "*"
pylint = "*"
rope = "*"
tox = "*"


[tool.poetry.extras]
server = [
    "django",
    "channels",
    "graphene-protector",
    "strawberry-graphql",
    "strawberry-django-plus",
    "django-fast-ratelimit",
    "django-manifest-loader",
]
proxy = [
    "bleach",
    "tinycss2"
]
fastcgi = ["flipflop"]
postgresql = ["psycopg"]
mysql = ["mysqlclient"]


[tool.black]
line-length = 79

[tool.isort]
profile="black"

[build-system]
requires = ["poetry_core"]
build-backend = "poetry.core.masonry.api"
